requirements.txt
# quart
# hypercorn
flask
flask_restful
waitress
colorama
mysql-connector-python
bcrypt